+++ FLASH6/flash6.js
@@ -20712,7 +20712,7 @@ function requestMobileMockup3dMesh(){
         fw_program:"Altis_Intelligent3_firmware1",
         fw_board:"Altis_Intelligent3_b3",
         fw_protocol:"Flash6-Intelligent-b4",
-        fw_build:"v6 b19"
+        fw_build:"v6 b20"
       };
     }